Add proper support for onReadConfiguration in device-specific settings
Only one device uses this function and it does so in a custom settings
activity and doesn't even use the config parameter...

Now, DeviceSpecificSettingsFrament will send a query to DeviceSupport
upon opening the configuration and the support has an option to retrieve
the actual configuration value from the device. It does so by first
indicating that it started the download by sending an IN_PROGRESS event,
and then sends a SUCCESS event once the setting has been retrieved
and updated inside SharedProperties.